backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.8",
    "pre-commit>=4.2.0",
    "psycopg[binary]>=3.2.4",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]

[tool.pytest.ini_options]
asyncio_mode="auto"
